+++ src/main/java/com/flab/gettoticket/repository/GoodsRepository.java
@@ -1,7 +1,7 @@
 package com.flab.gettoticket.repository;
 
-import com.flab.gettoticket.model.Goods;
-import com.flab.gettoticket.model.Zone;
+import com.flab.gettoticket.entity.Goods;
+import com.flab.gettoticket.entity.Zone;
 
 import java.util.List;
