+++ src/main/java/com/flab/gettoticket/service/PlayService.java
@@ -2,7 +2,7 @@ package com.flab.gettoticket.service;
 
 import com.flab.gettoticket.dto.SeatDTO;
 import com.flab.gettoticket.dto.SeatCountDTO;
-import com.flab.gettoticket.model.PlayTime;
+import com.flab.gettoticket.entity.PlayTime;
 
 import java.time.LocalDate;
 import java.util.List;
